Optimize, bugfix and resolve `clang-tidy` warnings. (#39)
* Formatting

* Optimize, bugfix and resolve `clang-tidy` warnings.

* Fix expanded code

+++ include/stdlib.h
@@ -13,6 +13,7 @@
 #define INCLUDE_STDLIB_H_
 
 #include "alloc.h"
+#include "stddef.h"
 #include "stdint.h"
 
 #define ZEROPAD 1  // pad with zero
@@ -23,11 +24,11 @@
 #define SPECIAL 32 // 0x
 #define SMALL   64 // use 'abcdef' instead of 'ABCDEF'
 
-#define do_div(n, base)                                                                 \
-    ({                                                                                  \
-        int64_t __res;                                                                  \
-        __asm__("divq %4" : "=a"(n), "=d"(__res) : "0"(n), "1"(0), "r"((int64_t)base)); \
-        __res;                                                                          \
+#define do_div(n, base)                                                                   \
+    ({                                                                                    \
+        int64_t __res;                                                                    \
+        __asm__("divq %4" : "=a"(n), "=d"(__res) : "0"(n), "1"(0), "r"((int64_t)(base))); \
+        __res;                                                                            \
     })
 
 /* Determine whether it is a number */
@@ -40,9 +41,9 @@ int atoi(const char *pstr);
 int skip_atoi(const char **s);
 
 /* Formatting an integer as a string */
-char *number(char *str, int64_t num, int base, int size, int precision, int type);
+char *number(char *str, size_t num, size_t base, size_t size, size_t precision, int type);
 
 /* Returns the size of a string with an integer formatted by `number()` */
-uint64_t number_length(int64_t num, int base, int size, int precision, int type);
+uint64_t number_length(size_t num, size_t base, size_t size, size_t precision, int type);
 
 #endif // INCLUDE_STDLIB_H_
