+++ kernel/mem/bitmap.c
@@ -10,7 +10,6 @@
  */
 
 #include "bitmap.h"
-#include "printk.h"
 #include "string.h"
 
 /* Initialize the memory bitmap */
@@ -30,7 +29,7 @@ int bitmap_get(const Bitmap *bitmap, size_t index)
 }
 
 /* Setting the memory bitmap */
-void bitmap_set(Bitmap *bitmap, size_t index, int value)
+void bitmap_set(Bitmap *bitmap, size_t index, int value) // NOLINT
 {
     size_t word_index = index / 8;
     size_t bit_index  = index % 8;
@@ -41,7 +40,7 @@ void bitmap_set(Bitmap *bitmap, size_t index, int value)
 }
 
 /* Set the memory bitmap range */
-void bitmap_set_range(Bitmap *bitmap, size_t start, size_t end, int value)
+void bitmap_set_range(Bitmap *bitmap, size_t start, size_t end, int value) // NOLINT
 {
     size_t start_word = (start + 7) / 8;
     size_t end_word   = end / 8;
@@ -56,7 +55,7 @@ void bitmap_set_range(Bitmap *bitmap, size_t start, size_t end, int value)
 }
 
 /* Memory bitmap search range */
-size_t bitmap_find_range(const Bitmap *bitmap, size_t length, int value)
+size_t bitmap_find_range(const Bitmap *bitmap, size_t length, int value) // NOLINT
 {
     size_t count = 0, start_index = 0;
     uint8_t byte_match = value ? (uint8_t)-1 : 0;
@@ -71,7 +70,7 @@ size_t bitmap_find_range(const Bitmap *bitmap, size_t length, int value)
             if (count >= length) return start_index;
         } else {
             for (size_t bit = 0; bit < 8; bit++) {
-                int bit_value = (byte >> bit) & 1;
+                int bit_value = (int)((byte >> bit) & 1);
                 if (bit_value == value) {
                     if (count == 0) start_index = byte_idx * 8 + bit;
                     count++;
